backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "cryptography>=46.0.1",
    "dependency-injector>=4.48.2",
    "fastapi[standard]>=0.117.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.36.0",
]

[project.optional-dependencies]
dev = [
    "alembic>=1.16.5",
    "mypy>=1.18.2",
    "ruff>=0.13.1",
]

[dependency-groups]
dev = [
    "alembic>=1.16.5",
]
